+++ src/client/contexts/GridStateContext.tsx
@@ -1,9 +1,13 @@
-import { createContext, memo, use, useState } from 'react';
+import { createContext, memo, use, useEffect, useState } from 'react';
 import { GridState, State } from '@logic-pad/core/data/primitives';
+import { GridValidator } from '@logic-pad/core/data/validateAsync';
+import { GridData } from '@logic-pad/core/index';
 
 interface GridStateContext {
   state: GridState;
   setState: (value: GridState) => void;
+  validateGrid: (grid: GridData, solution: GridData | null) => void;
+  gridValidator: GridValidator;
 }
 
 export const defaultState: GridState = {
@@ -15,6 +19,8 @@ export const defaultState: GridState = {
 const Context = createContext<GridStateContext>({
   state: defaultState,
   setState: () => {},
+  validateGrid: () => {},
+  gridValidator: new GridValidator(),
 });
 
 export const useGridState = () => {
@@ -37,11 +43,30 @@ export default memo(function GridStateContext({
   const state = externalState ?? internalState;
   const setState = setExternalState ?? setInternalState;
 
+  const [gridValidator, setGridValidator] = useState<GridValidator>(
+    () => new GridValidator()
+  );
+
+  useEffect(() => {
+    const validator = new GridValidator();
+    validator.subscribeToState(newState => setState(newState));
+    setGridValidator(validator);
+    return () => {
+      validator.delete();
+    };
+  }, [setState]);
+
+  const validateGrid = (grid: GridData, solution: GridData | null) => {
+    gridValidator.validateGrid(grid, solution);
+  };
+
   return (
     <Context
       value={{
         state,
         setState,
+        validateGrid,
+        gridValidator,
       }}
     >
       {children}